Include item entity in forms

+++ src/app/admin/api/content/update/route.ts
@@ -12,7 +12,8 @@ export async function POST(req: Request) {
 
     const formData = await req.formData();
     const label = formData.get("label") as Label;
-    const BDContent = await prisma.content.findUnique({
+
+    let BDContent = await prisma.content.findUnique({
       where: {
         label: label,
       },
@@ -25,19 +26,22 @@ export async function POST(req: Request) {
       },
     });
 
-    let content;
+    if (!BDContent) {
+      BDContent = await prisma.content.create({
+        data: {
+          label,
+          title: "",
+          text: formData.get("text") as string | "",
+          images: {},
+        },
+        include: { images: true },
+      });
+    }
+
+    const id = BDContent.id;
+
     // Content with only images (Label.SLIDER)
     if (label === Label.SLIDER) {
-      if (!BDContent) {
-        content = await prisma.content.create({
-          data: {
-            label: label,
-            title: "",
-            text: "",
-          },
-        });
-      } else content = BDContent;
-
       const files = formData.getAll("files") as File[];
       for (const file of files) {
         if (file.size > 0) {
@@ -46,7 +50,7 @@ export async function POST(req: Request) {
           const fileInfo = await resizeAndSaveImage(file, title, dir);
           if (fileInfo) {
             await prisma.content.update({
-              where: { id: content.id },
+              where: { id },
               data: {
                 images: {
                   create: {
@@ -62,37 +66,15 @@ export async function POST(req: Request) {
         }
       }
     } else {
-      if (!BDContent) {
-        content = await prisma.content.create({
-          data: {
-            label,
-            title: "",
-            text: formData.get("text") as string,
-            images: {},
-          },
-          include: { images: true },
-        });
-      } else {
-        content = await prisma.content.update({
-          where: {
-            id: BDContent.id,
-          },
-          data: {
-            text: formData.get("text") as string,
-          },
-          include: { images: true },
-        });
-      }
-
-      // Contents with one image (PRESENTATION)
-      if (label === Label.PRESENTATION) {
-        const file = formData.get("file") as File;
-        if (file.size > 0) {
-          if (content.images.length > 0) {
-            const oldFilename = content.images[0].filename;
+      const file = formData.get("file") as File;
+      if (label === Label.PRESENTATION && file) {
+        // Contents with only one image (Image PRESENTATION)
+        if (file && file.size > 0) {
+          if (BDContent.images.length > 0) {
+            const oldFilename = BDContent.images[0].filename;
             deleteFile(dir, oldFilename);
             await prisma.content.update({
-              where: { id: content.id },
+              where: { id },
               data: {
                 images: {
                   delete: { filename: oldFilename },
@@ -100,11 +82,10 @@ export async function POST(req: Request) {
               },
             });
           }
-
           const fileInfo = await resizeAndSaveImage(file, "presentation", dir);
           if (fileInfo) {
             await prisma.content.update({
-              where: { id: content.id },
+              where: { id },
               data: {
                 images: {
                   create: {
@@ -117,8 +98,19 @@ export async function POST(req: Request) {
             });
           }
         }
+      } else {
+        await prisma.content.update({
+          where: {
+            id,
+          },
+          data: {
+            text: formData.get("text") as string,
+          },
+          include: { images: true },
+        });
       }
     }
+
     return Response.json({ message: "ok" }, { status: 200 });
   } catch (e) {
     console.log(e);
